+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Implemented retries
 - Implement `logging`
 - Implemented download dir
+- Code split the solution
 
 ### Modified